# doc/api/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: N/A
  version: "1"
  title: Core API
host: localhost:3000
basePath: /api/v1
schemes:
- http
paths:
...
  /code/{id}:
    get:
      description: grabs code
      parameters: 
        - name: id
          in: path 
          type: string
          description: code id 
          required: true
        - in: query
          name: classCode
          type: string
          required: true
          description: classCode
      responses: 
        200:
          description: Return code 
          schema:
            $ref: '#/definitions/getCode'
  /code:
    post: 
      description: post code
      consumes: 
      - application/json
      parameters: 
        - name: record
          in: body
          required: true
          schema:
            $ref: '#/definitions/postCode'
        - in: query
          name: classCode
          type: string
          required: true
          description: classCode
      responses: 
        200: 
          description: post ok
          schema: 
            $ref: '#/definitions/postCode'          
definitions:
...
  getCode: 
     type: array
     items: 
        type: string
  postCode:
      type: object
      required: 
        - code
        - language
        - day
      properties:
        code:
          type: string
        language: 
          type: string
        day: 
          type: string